refined all the ui of all the pages, added option for updating and deleting the vehicle and enhance cards of prediciton
