+++ .streamlit/config.toml
@@ -1,5 +1,3 @@
 [server]
 runOnSave = true
 
-[client]
-showSidebarNavigation = false
